Abstracted Services.Configuration.ConfigurationSettings into appropriate namespaces
