--- src/Dfc.CourseDirectory.Services/Configuration/ConfigurationSettings.cs
@@ -1,42 +1,0 @@
-﻿namespace Dfc.CourseDirectory.Services.Configuration
-{
-    public class VenueNameComponentSettings
-    {
-        public string VenueName_Label { get; set; }
-    }
-
-    public class CourseForComponentSettings
-    {
-        public int TextFieldMaxChars { get; set; }
-    }
-
-    public class EntryRequirementsComponentSettings
-    {
-        public int TextFieldMaxChars { get; set; }
-    }
-
-    public class WhatWillLearnComponentSettings
-    {
-        public int TextFieldMaxChars { get; set; }
-    }
-
-    public class HowYouWillLearnComponentSettings
-    {
-        public int TextFieldMaxChars { get; set; }
-    }
-
-    public class WhatYouNeedComponentSettings
-    {
-        public int TextFieldMaxChars { get; set; }
-    }
-
-    public class HowAssessedComponentSettings
-    {
-        public int TextFieldMaxChars { get; set; }
-    }
-
-    public class WhereNextComponentSettings
-    {
-        public int TextFieldMaxChars { get; set; }
-    }
-}
