+++ src/Dfc.CourseDirectory.Services/CourseService/CourseForComponentSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfc.CourseDirectory.Services.CourseService
+{
+    public class CourseForComponentSettings
+    {
+        public int TextFieldMaxChars { get; set; }
+    }
+}
