+++ src/Dfc.CourseDirectory.Services/CourseService/EntryRequirementsComponentSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfc.CourseDirectory.Services.CourseService
+{
+    public class EntryRequirementsComponentSettings
+    {
+        public int TextFieldMaxChars { get; set; }
+    }
+}
