+++ src/Dfc.CourseDirectory.Services/CourseService/HowAssessedComponentSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfc.CourseDirectory.Services.CourseService
+{
+    public class HowAssessedComponentSettings
+    {
+        public int TextFieldMaxChars { get; set; }
+    }
+}
