+++ src/Dfc.CourseDirectory.Services/CourseService/HowYouWillLearnComponentSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfc.CourseDirectory.Services.CourseService
+{
+    public class HowYouWillLearnComponentSettings
+    {
+        public int TextFieldMaxChars { get; set; }
+    }
+}
