+++ src/Dfc.CourseDirectory.Services/CourseService/WhatWillLearnComponentSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfc.CourseDirectory.Services.CourseService
+{
+    public class WhatWillLearnComponentSettings
+    {
+        public int TextFieldMaxChars { get; set; }
+    }
+}
