+++ src/Dfc.CourseDirectory.Services/CourseService/WhatYouNeedComponentSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfc.CourseDirectory.Services.CourseService
+{
+    public class WhatYouNeedComponentSettings
+    {
+        public int TextFieldMaxChars { get; set; }
+    }
+}
