+++ src/Dfc.CourseDirectory.Services/CourseService/WhereNextComponentSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfc.CourseDirectory.Services.CourseService
+{
+    public class WhereNextComponentSettings
+    {
+        public int TextFieldMaxChars { get; set; }
+    }
+}
