+++ src/Dfc.CourseDirectory.Web/Configuration/VenueNameComponentSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfc.CourseDirectory.Web.Configuration
+{
+    public class VenueNameComponentSettings
+    {
+        public string VenueName_Label { get; set; }
+    }
+}
