+++ src/Dfc.CourseDirectory.Web/Startup.cs
@@ -10,7 +10,6 @@ using Dfc.CourseDirectory.Services.ApprenticeshipBulkUploadService;
 using Dfc.CourseDirectory.Services.ApprenticeshipService;
 using Dfc.CourseDirectory.Services.BlobStorageService;
 using Dfc.CourseDirectory.Services.BulkUploadService;
-using Dfc.CourseDirectory.Services.Configuration;
 using Dfc.CourseDirectory.Services.CourseService;
 using Dfc.CourseDirectory.Services.CourseTextService;
 using Dfc.CourseDirectory.Services.Models.Environment;
@@ -71,7 +70,6 @@ namespace Dfc.CourseDirectory.Web
 
             services.AddSingleton<IConfiguration>(Configuration);
             services.AddTransient<IUkrlpWcfClientFactory, UkrlpWcfClientFactory>();
-            services.Configure<VenueNameComponentSettings>(Configuration.GetSection("AppUISettings:VenueNameComponentSettings"));
             services.Configure<CourseForComponentSettings>(Configuration.GetSection("AppUISettings:CourseForComponentSettings"));
             services.Configure<EntryRequirementsComponentSettings>(Configuration.GetSection("AppUISettings:EntryRequirementsComponentSettings"));
             services.Configure<WhatWillLearnComponentSettings>(Configuration.GetSection("AppUISettings:WhatWillLearnComponentSettings"));
