+++ src/Dfc.CourseDirectory.Web/ViewComponents/Courses/HowYouWillLearn/Default.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Dfc.CourseDirectory.Web.Validation
-@using Dfc.CourseDirectory.Services.Configuration
+@using Dfc.CourseDirectory.Services.CourseService
 @using Microsoft.Extensions.Options
 @model Dfc.CourseDirectory.Web.ViewComponents.Courses.HowYouWillLearn.HowYouWillLearnModel
