+++ src/Dfc.CourseDirectory.Web/ViewComponents/VenueName/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using Dfc.CourseDirectory.Services.Configuration
+﻿@using Dfc.CourseDirectory.Web.Configuration
 @using Microsoft.Extensions.Options
 @model Dfc.CourseDirectory.Web.ViewComponents.VenueName.VenueNameModel
 
@@ -12,7 +12,7 @@
 
 
         <label class="govuk-label" asp-for="@Model.VenueName" id="VenueNameLabel">
-           Location name
+            Location name
         </label>
         @if (!string.IsNullOrWhiteSpace(Model.HintText))
         {
